选课获取全部课程添加日志(dev 2019-09-18) 005

See merge request tongji/course-election!1485

(cherry picked from commit d43ef2a5df1bd67fcafe3987925390aaac7368f5)

c14660bd 选课获取全部课程添加日志(dev 2019-09-18) 005

+++ src/main/java/com/server/edu/election/controller/ElecAgentYjsController.java
@@ -128,12 +128,12 @@ public class ElecAgentYjsController
         
         if (!Constants.PROJ_UNGRADUATE.equals(session.getCurrentManageDptId()))
         {
-            if (elecRequest.getChooseObj() == Constants.TOW)
+            if (elecRequest.getChooseObj().intValue() == Constants.TOW)
             { // 教务员
                 c = yjsService
                     .setData(studentId, c, elecRequest.getRoundId(), null);
             }
-            else if (elecRequest.getChooseObj() == Constants.THREE)
+            else if (elecRequest.getChooseObj().intValue() == Constants.THREE)
             { // 管理员
                 c = yjsService
                     .setData(studentId, c, null, elecRequest.getCalendarId());
